Upgrade ombi to v3!

+++ scripts/remove/ombi.sh
@@ -4,5 +4,13 @@ systemctl stop ombi
 rm /etc/systemd/system/ombi.service
 rm -f /etc/nginx/apps/ombi.conf
 service nginx reload
-rm -rf /opt/ombi
+
+if [[ -d /opt/ombi ]]; then
+  rm -rf /opt/ombi
+fi
+
+if [[ -d /opt/Ombi ]]; then
+  rm -rf /opt/Ombi
+fi
+
 rm /install/.ombi.lock
